requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
# bcrypt==4.0.1  # Commented out for testing
python-dotenv==1.0.0
PyMySQL==1.1.0
mysqlclient==2.2.0